fix!: move leaderboard and findteams tabs out, encode team names for getTeamInfo
